pyproject.toml
[tool.poetry]
name = "six-figure-rag-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
fastapi = ">=0.121.2,<0.122.0"
uvicorn = ">=0.38.0,<0.39.0"
supabase = ">=2.24.0,<3.0.0"
python-dotenv = ">=1.2.1,<2.0.0"
clerk-backend-api = ">=4.0.0,<5.0.0"
boto3 = ">=1.40.74,<2.0.0"
celery = ">=5.5.3,<6.0.0"
langchain = "1.0.8"
langchain-community = "0.4.1"
langchain-openai = "1.0.3"
unstructured = {version = "0.18.11", extras = ["all-docs"]}
scrapingbee = "^2.0.2"
redis = "^7.0.1"
ipykernel = "^7.1.0"
ddgs = "^9.9.1"
langchain-tavily = "^0.2.13"
ragas = "^0.4.3"
openai = "^2.30.0"
structlog = "^25.5.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"